Add simulation script to determine if scale in RFgauss follows range
The spatial landscape is not square so large values have multiple
correlations that may or not be correct because of the layout of the
landscape. Therefore, I chose to use a scale value of 1000 km.
The simulation is done with a population having a high density
because this leads to a reduced variance for the Poisson distribution.
Otherwise, the range parameter is not quite estimable.