Change the cpp file to use a recursive parameterization.
The use of a recursive paramaterization in the OM was
conflicting with the autoregressive paramaterization in the
EM and leading to biased paramater estimates. I belive the
EM is now fixed. Though I will save the old EM, under
spatial_gompertzautoregressive.cpp to keep the old EM
available without reverting commits.
More testing is needed.